improved v1 visualization
